requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.9
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
drf-spectacular==0.27.2
idna==3.11
numpy==2.1.3
pillow==12.0.0
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.2
gunicorn==23.0.0
psycopg2-binary==2.9.11
sentry-sdk==2.19.0
